SVM - rbf kernel
From chatbot openAI :

eps_SVR :
Epsilon-Support Vector Regression (Epsilon-SVR) is a variant of the Support Vector Regression (SVR) algorithm that uses a different approach to handle errors in the predicted values. While SVR uses a "tube" around the predicted function within which the true values must lie, Epsilon-SVR allows for a small amount of deviation from the predicted values by introducing a margin of error, or "epsilon", into the objective function.
The epsilon parameter in Epsilon-SVR determines the size of the margin of error, and it controls the trade-off between the simplicity of the predicted function and the accuracy of the predictions. A smaller epsilon value will result in a more complex and accurate function, but it may also increase the risk of overfitting to the training data. A larger epsilon value, on the other hand, will result in a simpler and less accurate function, but it will also be less susceptible to overfitting.
In summary, Epsilon-SVR is a variant of SVR that uses a margin of error, or epsilon, to allow for a small amount of deviation from the predicted values. This can help to improve the accuracy of the predictions without increasing the complexity of the predicted function, and it can also help to prevent overfitting to the training data.

RBF kernel :
A radial basis function (RBF) kernel is a type of kernel function that is commonly used in support vector machine (SVM) algorithms. A kernel function is a mathematical operation that is applied to the input data in order to transform it into a higher-dimensional space, where it is more easily separable. This allows the SVM algorithm to learn more complex and flexible functions that can accurately model the relationship between the input and output data.
RBF kernels are a type of "radial" kernel, meaning that they are based on the distance between the input data points. More specifically, an RBF kernel is defined as the Gaussian function of the Euclidean distance between two input points. This function has a bell-shaped curve that is centered at the origin, and it has a parameter, called sigma, that determines the width of the curve.
The RBF kernel is a popular choice for SVM algorithms because it can learn complex and non-linear functions that are well-suited to many types of data. It is also relatively easy to compute and implement, and it has a number of useful properties that make it well-suited for SVM applications.
In summary, an RBF kernel is a type of kernel function that is commonly used in SVM algorithms. It is based on the distance between the input data points, and it is defined as the Gaussian function of the Euclidean distance between two points. This allows the SVM algorithm to learn complex and non-linear functions that can accurately model the relationship between the input and output data.